test: make version check test-friendly (no import.meta, lazy load in tests)
