+++ src/services/versionTypes.ts
@@ -1,0 +1,8 @@
+export interface VersionInfo {
+  current: string;
+  latest: string;
+  hasUpdate: boolean;
+  whatsNew?: string;
+  url: string;
+}
+
